+++ SourceryRuntime/Sources/Equality.generated.swift
@@ -118,6 +118,15 @@ extension GenericType {
     override public func isEqual(_ object: Any?) -> Bool {
         guard let rhs = object as? GenericType else { return false }
         if self.name != rhs.name { return false }
+        if self.typeParameters != rhs.typeParameters { return false }
+        return true
+    }
+}
+extension GenericTypeParameter {
+    /// :nodoc:
+    override public func isEqual(_ object: Any?) -> Bool {
+        guard let rhs = object as? GenericTypeParameter else { return false }
+        if self.typeName != rhs.typeName { return false }
         return true
     }
 }